revert: restore original contributor wall
