+++ README.md
@@ -156,61 +156,61 @@ deepagents-in-action/
 ## 贡献者墙
 
 <!-- contributors:start -->
-<table align="center" width="100%">
+<table>
 <tr>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/webup" title="webup">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/webup">
     <img src="https://avatars.githubusercontent.com/u/2936504?v=4&s=144" width="72" height="72" alt="webup" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;&#160;&#160;webup&#160;&#160;&#160;</samp></sub>
+    <sub><strong>webup</strong></sub>
   </a><br />
   <sub>45 commits</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/knqiufan" title="knqiufan">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/knqiufan">
     <img src="https://avatars.githubusercontent.com/u/34114995?v=4&s=144" width="72" height="72" alt="knqiufan" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;knqiufan&#160;&#160;</samp></sub>
+    <sub><strong>knqiufan</strong></sub>
   </a><br />
   <sub>3 commits</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/Spr1ng7" title="Spr1ng7">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/Spr1ng7">
     <img src="https://avatars.githubusercontent.com/u/188573471?v=4&s=144" width="72" height="72" alt="Spr1ng7" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;&#160;Spr1ng7&#160;&#160;</samp></sub>
+    <sub><strong>Spr1ng7</strong></sub>
   </a><br />
   <sub>2 commits</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/1iyouzhen" title="1iyouzhen">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/1iyouzhen">
     <img src="https://avatars.githubusercontent.com/u/184539312?v=4&s=144" width="72" height="72" alt="1iyouzhen" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;1iyouzhen&#160;</samp></sub>
+    <sub><strong>1iyouzhen</strong></sub>
   </a><br />
   <sub>1 commit</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/codeMonkeyWang" title="codeMonkeyWang">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/codeMonkeyWang">
     <img src="https://avatars.githubusercontent.com/u/3906539?v=4&s=144" width="72" height="72" alt="codeMonkeyWang" style="border-radius:50%;" /><br />
-    <sub><samp>codeMonkey…</samp></sub>
+    <sub><strong>codeMonkeyWang</strong></sub>
   </a><br />
   <sub>1 commit</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/dongyu23" title="dongyu23">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/dongyu23">
     <img src="https://avatars.githubusercontent.com/u/101914323?v=4&s=144" width="72" height="72" alt="dongyu23" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;dongyu23&#160;&#160;</samp></sub>
+    <sub><strong>dongyu23</strong></sub>
   </a><br />
   <sub>1 commit</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/Walt-like" title="Walt-like">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/Walt-like">
     <img src="https://avatars.githubusercontent.com/u/56186222?v=4&s=144" width="72" height="72" alt="Walt-like" style="border-radius:50%;" /><br />
-    <sub><samp>&#160;Walt&#8209;like&#160;</samp></sub>
+    <sub><strong>Walt-like</strong></sub>
   </a><br />
   <sub>1 commit</sub>
 </td>
-<td align="center" valign="top" width="12.5%">
-  <a href="https://github.com/YangLeeFangfei" title="YangLeeFangfei">
+<td align="center" valign="top" width="120">
+  <a href="https://github.com/YangLeeFangfei">
     <img src="https://avatars.githubusercontent.com/u/111680955?v=4&s=144" width="72" height="72" alt="YangLeeFangfei" style="border-radius:50%;" /><br />
-    <sub><samp>YangLeeFan…</samp></sub>
+    <sub><strong>YangLeeFangfei</strong></sub>
   </a><br />
   <sub>1 commit</sub>
 </td>
