--- scripts/update-contributors.test.mjs
@@ -1,34 +1,0 @@
-import assert from 'node:assert/strict';
-import test from 'node:test';
-
-import { buildWall } from './contributor-wall.mjs';
-
-const contributors = [
-  {
-    login: 'codeMonkeyWang',
-    profileUrl: 'https://github.com/codeMonkeyWang',
-    avatarUrl: 'https://avatars.githubusercontent.com/u/3906539?v=4&s=144',
-    contributions: 1,
-  },
-  ...Array.from({ length: 7 }, (_, index) => ({
-    login: index === 6 ? 'Walt-like' : `user${index}`,
-    profileUrl: `https://github.com/${index === 6 ? 'Walt-like' : `user${index}`}`,
-    avatarUrl: `https://avatars.githubusercontent.com/u/${index}?v=4&s=144`,
-    contributions: 1,
-  })),
-];
-
-test('builds an eight-column table with clickable abbreviated names', () => {
-  const wall = buildWall(contributors);
-
-  assert.equal((wall.match(/<tr>/g) ?? []).length, 1);
-  assert.equal((wall.match(/<table align="center" width="100%">/g) ?? []).length, 1);
-  assert.equal((wall.match(/width="12.5%"/g) ?? []).length, contributors.length);
-  assert.match(
-    wall,
-    /<a href="https:\/\/github\.com\/codeMonkeyWang" title="codeMonkeyWang">\s+<img[^>]+><br \/>\s+<sub><samp>codeMonkey…<\/samp><\/sub>\s+<\/a>/,
-  );
-  assert.match(wall, /<sub><samp>&#160;&#160;&#160;user0&#160;&#160;&#160;<\/samp><\/sub>/);
-  assert.match(wall, /<sub><samp>&#160;Walt&#8209;like&#160;<\/samp><\/sub>/);
-  assert.doesNotMatch(wall, /<kbd>|<wbr>|<strong>|&#8203;|width="14.2857%"/);
-});
